requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
bidict==0.23.1
    # via python-socketio
dnspython==2.7.0
    # via eventlet
eventlet==0.39.1
    # via -r requirements.in
greenlet==3.1.1
    # via eventlet
h11==0.14.0
    # via wsproto
iniconfig==2.0.0
    # via pytest
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
pyside6==6.8.2.1
    # via -r requirements.in
pyside6-addons==6.8.2.1
    # via pyside6
pyside6-essentials==6.8.2.1
    # via
    #   pyside6
    #   pyside6-addons
pytest==8.3.5
    # via -r requirements.in
python-engineio==4.11.2
    # via python-socketio
python-socketio==5.12.1
    # via -r requirements.in
shiboken6==6.8.2.1
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
simple-websocket==1.1.0
    # via python-engineio
websocket-client==1.8.0
    # via -r requirements.in
wsproto==1.2.0
    # via simple-websocket

requests~=2.32.3